fix(reports): QSR paginated parity — legacy colors, teacher subtotal, drop invented variants, count semantics
Bring the Question Summary Report paginated family to visual + structural
parity with legacy GAINS, verified against the rendered SSRS sample PDFs in
historical-reports/ (Heat Sources "- color" / "- Teacher" and the Tell Me a
Story working notes).

- PAG-2 (colors): correct cell #99FF99, incorrect #FFCCFF, Score% 3-band
  #FFCCFF/<70 · #FFF591/70-80 · #99FF99/>=80. Added QSR_GREEN/QSR_PINK/
  QSR_YELLOW + qsrPerformanceColor() in colors.ts; left the softened PERF_*
  report tokens untouched. Dropped the IAD palette from the QSR cells.
- PAG-3: removed the KPI strip from the QSR page (legacy QSR has none).
- PAG-4 (variants): dropped the invented "Header Highlights" variant and
  the wrong single "Teacher Subtotal" row; implemented the real legacy
  "- Teacher" two-row per-instructor subtotal block (# Correct Answers count
  row + 3-band-colored Score % row). The broken-in-legacy "Teacher Subtotal"
  report is intentionally skipped.
- PAG-6 (counts): "# Correct Answers" is now the COUNT of fully-correct
  (green) cells, not a sum of points_received; Possible Points is the
  attempted-cell count and Score% = correct/attempted at every grain. Also
  fixed a dim_standard alias-row fan-out in the matrix SQL that inflated the
  counts. Athenian item 8359960427 grand total now reads 308/486 = 63.4%,
  reconciling exactly with count-of-green cells (was 318 summed points / 323
  fanned).
- PAG-9 (header): legacy header order — title, assessment type, then
  "<Subject> - <Grade>: <Item_Name>" with no "Course:" prefix; removed the
  school logo (legacy QSR PDFs have none).
- PAG-8 (nav): QSR + paginated QRA routes are now reachable from the
  in-app ReportTypeSwitcher, not URL-only.
- PAG-5 (redacted, Decision 5): added a redacted variant that anonymizes
  names client-side (deterministic per uid). Exact legacy-hash parity via
  cube_user_summary.*_name_hash is a documented follow-up (no sample
  redacted PDF exists to diff against).

+++ backend/app/schemas/reports.py
@@ -602,8 +602,10 @@ class QsmStudentRow(BaseModel):
     user_uid: str
     user_name: str
     score_pct: float
-    possible_points: float
-    correct_count: float
+    # Count-based (PAG-6): possible_points = attempted-cell count,
+    # correct_count = count of fully-correct (green) cells.
+    possible_points: int
+    correct_count: int
     cells: dict[str, int | None]
 
 
@@ -614,11 +616,13 @@ class QsmTeacherGroup(BaseModel):
 
 
 class QsmGrandTotal(BaseModel):
-    possible_points: float
-    correct_count: float
+    # Count-based (PAG-6): possible_points / correct_count are cell counts,
+    # not summed points; per_question_* mirror the legacy footer rows.
+    possible_points: int
+    correct_count: int
     score_pct: float
-    per_question_possible: dict[str, float]
-    per_question_correct: dict[str, float]
+    per_question_possible: dict[str, int]
+    per_question_correct: dict[str, int]
     per_question_pct: dict[str, float]
